Refactor package of shardingsphere-proxy-frontend-core
